moves relavent tests from general admin to document view
